Update MultiDataProcessing with corrections
